Refactor: Move root command to a separate file
This commit begins the process of refactoring the CLI by moving the root command to its own file.
